+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Upgrade dependencies ([#267](https://github.com/interactivethings/catalog/pull/267))
 - Enable imports in all Specimens ([#154](https://github.com/interactivethings/catalog/pull/154))
+- Fix weird scale behaviour of the thumbnail in download specimen [#260](https://github.com/interactivethings/catalog/issues/260)
+- Create a new stacking context for the whole <Page> component [#223](https://github.com/interactivethings/catalog/issues/223)
 
 ## 2.5.3